perf(webapp): componetization pieChart
